Someone just got 4246 warnings on one page like
  Warning: object "EN HIV PUB STI STD CONTROL 1998" had no associated
  workflow. It has been assigned to the "aWebteam" workflow.

after they clicked Save from a story and returned to their Workspace.
I believe it is because for some reason we have thousands of stories
all in one desk. I think they got there because of a bug in bric_soap
but I didn't track that down yet. For some reason, these 12000 or so
stories and media have workflow__id == 0 and desk__id == 0 (presumably
after being published or shelved) but they still don't have a non-zero
grp__id in their `member' row (through story_member). They all have the
grp__id from the asset_grp number in the `desk' table, which is the
ID for the start desk of the first workflow. So this causes them to
all be listed when someone looks at a desk/workspace, apparently,
when you use ->list({grp_id => $foo}) (somewhere, I can't remember
where that was now..). So when you get to this line in the code,
you've got a story with $work_id == 0, so it calls $put_into_wf
to force it into the start desk of the first workflow. So, since
workflow__id == 0 is also valid (Shelved workflow), I thought I'd try
changing this to defined($work_id) instead. It's possibly wrong.