Fix for Bug # 1038. Media files are now stored in directories with media IDs broken into two-character strings. Version directories now have "v." prepended to them to disambiguate them from ID directory parts.
So, for example, if you had a media document "foo.txt" with the ID 1024 with
versions 0-3, where previously you had:

  /1024/0/foo.txt
  /1024/1/foo.txt
  /1024/2/foo.txt
  /1024/3/foo.txt

Now you have:

  /10/24/v.0/foo.txt
  /10/24/v.1/foo.txt
  /10/24/v.2/foo.txt
  /10/24/v.3/foo.txt

And if you had a media document "bar.txt" with the ID 10242, you will now have:

  /10/24/2/v.0/bar.txt
  /10/24/2/v.1/bar.txt
  /10/24/2/v.2/bar.txt

So now the largest number of subdirectories possible in any directory is the
number of versions of a media document plus 100 (0-99). Let us hope that no
one ever creates 30,998 versions of a single media document--at least not
before they upgrade their file system!