Added file upload and download fields for templates. In the process, I fixed the the JavaScript selectText() function so that it actually works with Mozilla and IE. Well, theoretically it works with IE. I found it to be a bit buggy :-(. I also discovered that Safari supports the Mozilla setSelectionRange method, which is sweet!
What else? I fixed the Media profile download section so that it uses JS to
launch the download Window. This is the first step towards making this section
XHTML/CSS. I'll get to the rest later.

I moved the find dialog and cheat sheet html in the template profile to after
the code textarea box, to fix errors on IE and to allow a field to be properly
highlighted when that interface loads.

That's it. Bug # 1150 is the new upload/download stuff for templates.