Removed the <iframe> for the left nav, replacing it by Javascript + cookies + CSS. Whenever you click on one of the menus, it updates a cookie like
  BRICOLAGE_MENUS=admin:open+adminSys:closed+1024:open+1025:closed

where you have numbers for workflow IDs, plus 'admin', 'adminSys',
'adminPub', and 'adminDist' which are the ids of the <li> in the
ADMIN menu.
  The menu toggling is done in a javascript callback with the function
toggleMenu in comp/media/js/lib.js. This does two things: changes the
classname of the <li> for whatever menu you're toggling, and updates
the cookie. The cookie updating part is a little ugly.
  The Nav callback class is gone now. The code in sharky/sideNav.mc
could use some refactoring.