Added in-place addition of top-level subelements.  By "top-level" I mean that you can currently only add subelements to the main story/media container, and not recursively through all of the subcontainers (yet).
This is the first use of partials in Bricolage.  The new add_element.html
creates the proper subelement and adds it to the current container, then
returns HTML for the new subelement to be inserted into the page via AJAX.

I spent way longer than I would have liked struggling with HTML::Prototype.
I'd really like to be able to use it, but it's too restrictive (link_to_remote
can only generate anchors, for instance).  Perhaps it can be subclassed to
override what I need.

NOTE: Backwards-incompatible API change:  
Bric::Biz::Element::Container::add_field() now returns the added field instead
of the parent container ($self).  This behavior matches that of add_container()
and is needed for rendering the partials.