Small change of plans.  Now instead of simply adding/removing an element from the DOM to add/delete something, we refresh its parent container.  This will be a bit slower, but it provides a few advantages: 1) The 'Add Element' list stays in sync; 2) the server-side state stays in sync; 3) the same partial can be used when adding an element, deleting an element, or expanding subcontainers.
The next step is to limit recursion to a reasonable level, probably 2. Any
containers deeper than that will appear collapsed.  When you click an arrow,
there will be an AJAX request which queries the database for that container
only and expands it.  In this way, we don't risk degrading performance on
really long stories.