Added autocompletion to keywords in the story profile.  As you type, a list of existing keywords appears, letting you choose from one of them, or type your own to create a new one.
The autocompletion code is almost generic enough to be used for things other
than keywords.  I intend to fix that, but I want to commit a working version
before I break anything.

Also fixed a bug that caused a newly-added keyword that belongs at the end of
the list to be added to the front instead.

TODO: People without permissions to create new keywords should not be allowed
to add keywords that don't exist.  Similar to categories.