AJAX desks!
Most of the work necessary to AJAXify the desks is done.  Most notably, you can
now delete, check in, check out, and move assets via AJAX.  The select list to
choose which desk to move/check in to has been replaced by a button with a
popout menu.  When you click the desk name, the story is checked in immediately
and the desk item disappears.

Previously, you couldn't move items while in My Workspace, and you couldn't check
in to a different desk while on a workflow desk.  Now you can.  In fact, there
are no visual differences between the workflow desks and the workspace, and the
workspace callbacks have been eliminated (merged with desk callbacks).

Icons are used for all of the actions (edit, notes, clone, trail, delete, 
checkin, move, checkout, etc.), except for log (will fix).  These icons are
from the Tango icon set (http://tango.freedesktop.org).

I still need to fix publishing.  The "Publish" checkbox is still there on
publish desks, even though the button to "Publish Assets" is gone, so you can't
actually publish anything.  I also added a "and Publish" option to the "Check
in to..." options, but it doesn't work yet either.  The icons also need to be
positioned better.