Fixed image document reversion as reported in Bug #1352. Turns out that, although the `revert` method was caching the updated media document in `pnotes()`, it was doing so with the Image subclass. The request, meanwhile, was looking up the media document as an "Media" class, and finding instead of the reverted object the media document reverted to. This royally screwed things up. Bottom line is that media documents should never be cached by `cache_me` as a subclass of the Media class.
So this commit adds a new method, `cache_as`, which defines the package for
which an object should be cached. By default it just returns `ref $self`, but
Media overrides it to always return Media. This way, subclasses of Media that
cache themselves will do so properly. And thus, reversion works, because the
reverted to media document looked up and cached is replaced when the document
that reverts to that looked up document reverts itself and replaces the
reverted to document in the request cache.

We've had some other issues in the past with subclasses of Media confusing
things in the cache. This should prevent them from happening again.