Fixed a bug where passing `version_id` to `lookup()` for a document or template will fail if the version ID specified is the same version number as the currently-checked-out version. This made publishes break, as noted in Bug #1417. I get around it by eliminating the code that checks for whether a version is checked-in or checked-out, the same as the `published_version` parameter does. Tests were added to replicate the issue before I fixed it.
Also, I changed the `checkin()` method to increment `current_version` instead of `version`, hoping this will fix the bugs we had with bogus version numbering. I'm going to revert that in a bit and write some more tests to replicate it, then make sure that's really the correct fix. Later tonight, I expect.

I also removed the stuff added in r8408, as that was not the correct fix to Bug #1417 (it ''really'' broke the publish lookups!).