Fixed a bug where story version numbers can get screwed up when a past version of an asset is published while newer versions are being edited. What I've done is prevent the `current_version` column from ever being updated unless the document is currently being checked in.
I'm not entirely convinced that this completely fixes the problem, because it only seems to come up when when a new version is created while an older one is in the process of being published. I suppose that it can happen for long publishes, and it definitely is a race condition that I was able to replicate in the tests and then fix, so maybe it will eliminate the problem. Keeping my fingers crossed!