A bit more precise fix for Bug # 93
The previous fix (commit 777112) was a bit brute-force, in that it could
delete completely irrelevant data from the session. Whilt it was probably
harmless, I updated it to first check the asset version before removing it
from the session. I also updated the `checkout()` callback in
`Bric::App::Callback::Desk` to clear the asset cache from the session, since
on checkout it needs to be completely reloaded no matter what. Each of these
changes addresses the exhibited bug on its own, but I thought it safer to be a
bit more thorough and precise, in the hopes that we can eliminate some other
cases where users end up with screwed-up documents.