Refactor mark_as_published()
I updated the docs for the new `mark_as_published()` method in
`Bric::Biz::Asset::Business`, removing some invalid notes, adding some valid
notes, and fixing some grammar nits. I also removed the call to `save()` at
the end of the method, because no method should call `save()`; the user should
always call it implicitly. I also changed the method so that it simply returns
if the document is already marked as published, as there is no point in
updating the publish date in this context. And finally, I added some tests
for the new method.