Fixed a bug in find_or_create_alternate()
Fixed a bug in `find_or_create_alternate()` where an element type is passed
with a different setting for its `top_level` attribute than is the value of
the `top_level` attribute of the original image's element type. This caused
problems because it looked for an existing image using the wrong URI. So I
hacked in a workaround to fool the current image into using the alternat's
element type when determining the URI, so that the URI should always be the
same. I ran into this in production for The Tyee, where editors were trying to
create non-fixed thumbnails for fixed images. This caused some pretty serious
indigestion. The workaround is a bit of a hack, but it's clean and works well.
I updated the test suite to replicate the issue before patching, so we do have
a proper regression test for the issue.