Exploit DBI callbacks to make MySQL behave.
Fixed compatibility issues on MySQL. Bricolage now sets `sql_mode` when it
connects to the database, so that MySQL itself does not have to globally set
that option, as it could cause conflicts with other apps running on MySQL.
Also removed the attempt to set C<sql_mode> globally at installation time; it
was bad manners, really.

I used the DBI's callback support to make this work, so bumped up the minimum
required DBI to 1.49. That's five years old already, so everyone should be
fine.