Contributors can be deleted again.
Looks like there was a thinko in 8a7d34b1652eba7c288cf35681b44cd37cdc85f6 such
that a select field was using the name for the id even when an `$id` argument
was passed. So I've fixed it to use the `$id` and to fallback on the name when
the `$id` isn't set. [#111 state:resolved].

There seems to be an unrelated issue where submitting a search in the
contributor popup by hitting "enter" instead of the "search" button causes any
contributors already added to be forgotten. What happens is the "enter"
submits the contributor's "Delete" button instead of the "Search" buttton.
This is really irritating. I'm not sure how to fix it, though; I'll
create a new bug to track it.