Fix failing MySQL tests
I tried for hours to reduce this to a simple regression, and failed. The
problem is that, during the tests, sometimes MySQL doesn't consider a date
with .00000 microseconds to be the same as without the .000000 when used in
the first part of a BETWEEN query. No idea WTF that's about, so work around it
by simply dropping the first date back a second in the tests.