Comment out `AddModule` in `httpd.conf`.
The `AddModule` lines are now commented-out in the `httpd.conf` generated by
the installer. As near as I can tell from my research, this directive hasn't
been relevant for a long time. Since we require 1.3.34 (ca. 2005), It seems
good enough to comment it out and add a note to uncomment them if necessary.

Backported from master.