Fix element type pagination bug
Paginated results in the element type manager no longer lose top-level element
type results when clicking the links for subsequent pages. Fixed by preventing
the `list()` method of Bric::Biz::ElementType from modifying the parameter
hash passed to it, as the modification was persisting in the list manager's
state. Thanks to Aaron Fuleki for the report and for making the Denison
available to diagnose the problem. [#102 state:resolved]