Avoid `//` in distribution directory names
This came up as a bug for a client: the FTP distributor was failing when the
destination file name was `/path/to//bricolage/uri/here`. So we remove the
leading `/` from Bricolage URIs to prevent the problem. Not sure if it applies
to SFTP or WebDAV, but it seems prudent to provide the same fix for them.

We could really use some tests for this stuff. Any volunteers?

Backported from master.