Don't return error on invalid category.
The category autocompetion feature in the "New Story" profile now refuses to
submit if a non-existant category has been typed in. If the user somehow
manages to submit anyway, she'll be returned to the form with a message asking
for a proper category name, rather than getting an error. Thanks to Scott for
the report. [#78 state:resolved]