Prevent `Category->list` from modifying params.
As with the bug with element types fixed a few days ago, this was causing
problems for the search interface, where params are stored in the session
cache. If the API is changing the values this mucks things up for subsequent
searches. This manifested in the site column appearing and dissappearing in
the category list search results. Backported from master, and tests added
and a bug thus found fixed. Will patch master with that fix, too.