Improve contrib popup button placement.
Moved the "Cancel" and "Save" buttons in the contributor popup up from the
bottom of the screen (it was below search results) to under the list of
contributors currently associated with the story and above the search box.
This makes it much easier to find the buttons when there are a lot of search
results. Thanks to Scott for the report [#84 state:resolved]