Handle character encoding in the VFTP server.
The virtual FTP server now assumes that template files uploaded to it are
encoded in the user's preferred encoding. So if you have Windows gremlins in
your templates, you can set your Character Set preference to "CP1252" and it
will just work. Thanks to Phillip Smith for the report [#124 state:resolved].