Add reversion to desk view.
Added the "and Revert" or "and Delete" option to the "Check In" menu on desk/My Workspace views. This saves the user from having to click "Edit" and then "Cancel Checkout". Suggested by Matt Rolf and Phillip Smith. In order to make this happen, the code that handles cancelling a checkout, which was duplicated nearly verbatim in the callbacks for the story, media, and template profiles has been factored out into a new class `Bric::App::Callback::Util::Asset`. [#62 state:resolved]