Update desk, workspace, and asset help.
Went to update the context-sensitive help to reflect yesterday's addition of the "and Revert" option in the "Check In" select lists, and found it pretty wanting. So I fixed it up for desks, My Workspace, and the main page for editing stories, media, and templates.

While I was at it, I changed the "Trail/Log" label on desk items to simply "Log".