Disable `UseCanonicalName`
It turns out that, with this on, a request for a directory sans slash would cause Apache to redirect to the domain name specified in the `ServerName` directive. This was annoying if the request was sent to a different host name. Such was the case, I noticed, on my Mac, where the canonical name was set to "benedict.local" but I just used "localhost" in my browser. I'd click the "Cancel" link in a destination action profile and be confronted with the login page. This is because that link requested `/admin/profile/dest?id=1234` and mod_dir would redirect it to `/admin/profile/dest/?id=1234`, and in that redirect, it would use the canonical server name rather than the requested server name. By turning off `UseCanonicalName`, we get mod_dir to redirect to the requested host name, instead, thus avoiding that particular issue.