Fix workflow forgetting bug.
Bric::Util::Burner's `publish()` method no longer removes documents from workflow after publishing them, as such is the responsbility of the code that calls publish. The publish callback already did, and now so does the Bric::SOAP::Workflow `publish()` method. Thanks to David Beaudet for the report and Aaron Fuleki for the steps to replicate.