Switch to Bricolage user after writing PID file.
Found that when Bricolage was configured to write the `bric_queued` and
`bric_ftpd` PID files to standard system locations that it would get
permission errors because, even though they were started by root, they were
becoming the Bricolage user before doing much else. This commit makes it so
that they become the Bricolage system user only after writing out the PID
file.

While at it, I changed the default location of the FTP server PID and log
files to the `log` subdirectory, rather than the bricolage root itself. The
log directory is clearly the better location, and has no permissions issues.
It's also what was always suggested by the comments in `bricolage.conf`,
though not in Bric::Admin or the code.