Allow autosuggest mininum character control.
Add the `CATEGORY_MIN_CHARS` AND `CATEGORY_MIN_CHARS_TMPL` `bricolage.conf`
directives. These give adminstrators the ability to control the responsiveness
of the autosuggest box for selecting categories to associate with assets. It
has been set to 2 to make it more responsive to sites with very large numbers
of categories, but most folks only need it set to 1. [#142 state:resolved]