Replace `getElementsByClassName()` with `select()`.
Adrian Yee pointed out that `getElementsByClassName()` will find all elements
with that class in the entire document, whereas calling Prototype's `select()`
with a class name selector on a single element will find only elements with
that class name under the element. This is more efficient, and much more
likely to be correct.