Respect publish_status in the SOAP API.
When creating new stories or media via the SOAP interface, no longer force the
publish status to be set to false, but set it to true if the the publish date
is set or the publish status is true in the XML. In such cases, also set the
published version number to version 1. This is to make it easier for one to
migrate documents from one Bricolage instance to another without having to
edit the XML to set the publish status to a true value.

While at it, fix a bug where the SOAP interface was not setting the published
version to the new version of a story or media document that is being updated
by XML that says the document is published and where the published version was
not previously set.