Guess at a fix for Bug #139.
I still can't replicate it. So as a guess, I'm making sure to declare the
`subsysTabs` JavaScript variable before it gets used. I kinda doubt that's it,
but it was the best I could come up with.

If that doesn't work someone else likely will have to fix it -- someone who
can replicate it.

While at it, I saw there was an `<input>` inside an `<h3>`, which was ugly, so
I fixed that. I also tweaked a regular expression to be more efficient.