Fix root category autocomplete bug.
Fix subtle category autocomplete issue where it can be difficult to choose the
root category. It would appear to work, but the "Unable to add category that
does not exist" message would not clear and the form would not think that the
category is legal. Could apply to other categories, but was most common for
the root category. It turns out it happens if the category has been fetched by
a previous request, and thus is cached. When the JavaScript code sees that
it's cached, it just needs to be sure to call the callback to indicate that
results were found. Thanks to Michael Herring for the detailed steps to
reproduce [#157 state:resolved]. [David]