Don't delete preview files on publish.
This happened for stories in multiple categories -- and probably for stories
with multiple pages, too. The problem was that `not_uri => ANY(@files)` was
doing `(uri NOT LIKE $file OR uri NOT LIKE $file)`. That `OR` need to be an
`AND`. But I realized that the `NONE` function works better. So now it's `uri
=> NONE(@file)`, which becomes `NOT( uri LIKE $file OR uri LIKE $file)`, which
is more correct.

Now to figure out how to backport, since there's no `NONE` in 1.10.