Don't delete preview files on publish.
This happened for stories in multiple categories -- and probably for stories
with multiple pages, too. The problem was that `not_uri => ANY(@files)` was
doing `(uri NOT LIKE $file OR uri NOT LIKE $file)`. That `OR` need to be an
`AND`. So I fixed the `any_where` function to detect `NOT LIKE` and switch
to an `AND` in that case.

(Just as an aside, I hate that we're parsing SQL bits here, but at least
they're all bits we control in our code).

Thanks to Guido Buelskaemper for the report [#156 state:resolved].