Port `NOT LIKE` fix from rev_1_10.
I already fixed the preview deletion issue in
631db647ae41d0a41150e97c27d0a4331600692e, but that option wasn't available in
rev_1_10. While poking there, I realized that the correct fix was to fix
`any_where` to recognize "NOT LIKE" and to switch to `AND`ing items instead of
`OR`ing them in that case. This matches what it was already doing for `<>` and
`!=`. So it makes sense to fix that issue here, too.