Refactor Bric::Util::Trans::SFTP.
It was a bit of a mess. Much of the code was unnecessary, dating from before
2.0 when it used Net::SFTP instead of Net::SSH2. As a result, the error
handling was all wrong.

While at it, remove the file name escaping, as with newer versions of
Net::SSH2 at least (or perhaps newer versions of libssh2), it was leaving the
backslashes in the file names on the destination server. So bump up the
minimum version of Net::SSH2.

For sanity, also add tests to make sure that Net::Util::Trans::SFTP and
Net::Util::Trans::SFTP compile and have apprpropriate interfaces.