Don't deactivate assets while updating via SOAP.
I've no idea why they were able temporarily disabled to begin with. Seemed an
unnecessary defense against something or other. And the downside was
discovered with the addition of the `EXPIRE_ON_DEACTIVATE` directive: assets
were getting expired every time they were updated via SOAP! The feature is in
2.0, but it seemed prudent to remove the unnecessary deactivation from 1.10,
too. So change the code to only activate or deactivate if requested.