Add `--priority` to `bric_soap` and `bric_republish`.
This allows bulk publishes to be lower priority, so that active publishes by
users can take precedence.

To make it consistent, also add a `$priority` agument to the `publish` method
of Bric::Util::Burner, and have Bric::Util::Job::Pub pass in its priority.
This is so that the burner can use it when it creates the distribution job
that corresponds to the publish job.