Don't sleep if there are more jobs to run.
Modify `bric_queued` so that it only sleeps for `--delay` seconds if it
executed no jobs. Otherwise it immediately runs again, on the assumption that
there will be more jobs that should be executed sooner rather than later. The
upshot is that files will now be distributed much sooner after publishling.