Use currently-checked out asset for diff.
If the "to" choice is the same version as the currently-checked out version,
use the currently-checked out representation. This is so we can see changes
since the checkout.

This is not ideal, frankly, as it's a bit magical. We should consider
modifying the functionality for 2.2 so that there is another choice for
currently-checked out items, "Current version" or something. When that's
selected, the currently-checked out version should be used (for "to" or
"from"). Otherwise, use the checked-in version. I think this would be less
surprising.

Conflicts:

	lib/Bric/Changes.pod