Don't update publish status on publish failure.
Or, more specifically, revert it to what it was before. If a story hasn't been
published before, then on publish failure the publish status will be false and
the publish date and first publishd date will be `undef`. If it *has* been
published before, then the publish date will not be udpated. This has been an
issue for a long time, so I'll backpatch all the way to 1.10.x.