Fix Xinha.
Update the JavaScript code to only load the editors when they're made visible.
This is a little fragile, since if we change the structure of the HTML at all,
we'll need to make sure that the Xinha code is updated to take advantage of
the new structure. That is, CSS selectors passed to `Element.select()` would
need to be updated.

Not sure if htmlArea or FCKEditor need to be updated, as well. [#217 state:resolved].