Restore category recall to new asset pages.
Before we had autocompletion of categories in the New Story/Media/Template
profiles, it used to be that they would remember the last category selected
when you go to create a new document. That was lost; so restore it.

Done by adding a new widget, `comp/widgets/profile/autocomplete.mc`, which
munges the memorization code from the `select_object` widget with the simple
population of a text filed in `text.mc`. Also a callback that handles
remembering the value when a form is submitted. That callback, of course, is
Bric::App::Callback::Autopopulate.

[#159 state:resolved]