Move related to their own twisty show/reveal boxes.
They're now also below container element fields, rather than above. They are
not draggable. When the related twisty is open, you see the thumbnail, edit
icon, and metadata. When it's closed, the metadata is hidden and the thumbnail
shrinks to 40px high and moves up. So the thumbnail always shows.

Still to do:

* Show related title, doc type when twisty is closed.
* Have edit icon open twisty *and* open edit dialog when it's tapped and the
  twisty is closed.
* Audit the metadata displayed when the twisty is open (move UUID to a
  tooltip).