Show two lines of context for closed relateds.
Just duping the first two lines of the summary: A definition list with the
related document title and element type. It gets toggled on whenever the
longer summary is toggled closed and vice versa. It looks pretty good in the
finished state, but jumps around a bit during the animation, but I'm not sure
it's worth worrying about.

In passing, localize the "ID" and "UUID" labels in the `title` attribute of
the summary `<dl>` element, and remove an unneeded JavaScript variable.