Show related info in element closed state.
So if an element is related story and/or media, show the titles of the
relateds below the hint (if there is a hint).

To keep things consistent and to adhere closer to the new design spec, change
the layout of the short summary for a related box to use a hint paragraph
instead of a definition list.