Add `assets_count()` to Desk and use it for desk counts.
This is a class method that returns a hash reference mapping desk IDs to counts of assets on those desks. Desks not in the hash refernce should be considered to have 0 assets.

Using it to provide desk counts when the "Show Desk Asset Counts" preference is on. This should be *far* more efficient than the old approach of loading every asset on every desk. [#239 state:resolved]