Fix drag and drop on IE9.
This is a direct hack to Prototype. The real fix is in Prototype 1.7, but I
thought it best not to throw that into the mix for a minor release. So I
updated it (and script.taculo.us) in master for 2.1.0, but just added a simple
hack for 2.0.1, borrowing code from [this bug report](https://prototype.lighthouseapp.com/projects/8886/tickets/1229-eventisleftbutton-failure-in-ie9).

[#250 state:resolved]